Split rtm tests into separate categories.
Add more tests around disconnect handling.
